Move functions to utils.ts
