+++ __tests__/main.test.ts
@@ -1,10 +1,4 @@
-import {
-  extractLabels,
-  getName,
-  getChecked,
-  formatStrArray,
-  validateEnum,
-} from '../src/main';
+import { extractLabels, getName, getChecked } from '../src/main';
 
 describe('main', (): void => {
   it(extractLabels.name, (): void => {
@@ -33,26 +27,5 @@ describe('main', (): void => {
 
   it(getChecked.name, (): void => {
     expect(getChecked({ checked: true })).toEqual(true);
-  });
-
-  it(formatStrArray.name, () => {
-    expect(formatStrArray(['a', 'b', 'c'])).toEqual('- a\n- b\n- c\n');
-    expect(formatStrArray([])).toEqual('');
-  });
-
-  it(validateEnum.name, () => {
-    enum B {
-      B = 'b',
-    }
-    expect(validateEnum('a', 'b' as string, B)).toBeUndefined();
-
-    enum CD {
-      C = 'c',
-      D = 'd',
-    }
-    const f = (): void => {
-      validateEnum('a', 'b' as string, CD);
-    };
-    expect(f).toThrow(new Error("`a` must be one of ['c', 'd'], but got 'b'"));
   });
 });
